Working Health And Mana inventory Pick up
Working Health And Mana inventory Pick up
To scroll the this inventory I set the buttons to "R" for up ,"F" for down in list, and "E" to useItem. When you use the item it will remove it from the list and the selectedItemPos on screen.

I still have to add Prefabs for HealthBoost and ManaBoost.
And I need to Make this Inventory Toggle on/off by player control.
Fixed Enemy script to stop taking Freeze and burn effects after they were destoryed because it kept throwing a null ref.