PlayerController and Scene transition
Ok I fixed issue with double spawn player between scenes with DontDestroyOnLoad() but deleting object in entry but  the player is still spawning in the same location as first scene and falling in void. I even made AreaEnterence Prefab and script to pull the players destination like teleporter but it is not working. I think my brain Is just fried at this point.